fix #112: update state on debugger start
After connecting to the VM the plugin requests the initial runtimestate but did not update the bridge with that state.
Consequently, the debugger faultly indicates the first line of any source file as the starting debugging location.
This commit updates the state of the bridge on connection and highlights the line associated to the PC retrieved from the first state as starting debug location.