Use JSON local index as index for the locals in source map
Now that the wat2wasm got changed to print a json that includes per local its index in commit https://github.com/TOPLLab/wabt/commit/89d092baf7581caa0908a18ac86377929443a952, we can use this information to keep track of the index of each local.
This way we can properly distinguish when a local is an argument or not.
We use this information to reimplement the `getLocals` to only use the localVariables that correspond to locals and not parameters.