display the name of the argument in de gui if applicable
Before this commit, the argument values would appear in the `VARIABLES` view of the plugin as `arg1`, `arg2`, etc.
Now we use the name of the argument only if possible. If the argument has no name we fallback to the default name `argi` where `i` is the number of the argument.